refactor: entities -> features 이관
