--- src/entities/post/ui/index.ts
@@ -1,2 +1,0 @@
-export * from "./PostHeader"
-export * from "./PostTable"
