+++ src/features/post/ui/index.ts
@@ -1,0 +1,4 @@
+export * from "./PostHeader"
+export * from "./PostTable"
+export * from "./PostDialogs"
+export * from "./PostSearch"
